Place Queens on the Chess Board
N-Queens Problem: Given a chess board having  cells, we need to place  queens in such a way that no queen is attacked by any other queen. A queen can attack horizontally, vertically and diagonally.

So initially we are having  unattacked cells where we need to place  queens. Let's place the first queen at a cell , so now the number of unattacked cells is reduced, and number of queens to be placed is . Place the next queen at some unattacked cell. This again reduces the number of unattacked cells and number of queens to be placed becomes . Continue doing this, as long as following conditions hold.

The number of unattacked cells is not .
The number of queens to be placed is not .